requirements.txt
telethon
binance-connector
